+++ src/Engine/OpenGL4Advance.h
@@ -2,13 +2,34 @@
 
 #include "OpenGL4Simple.h"
 
-namespace gl4A
+namespace gl4
 {
 	class ScopedDebugMarker final
 	{
 	public:
 		ScopedDebugMarker(const char* message);
 		ScopedDebugMarker(const ScopedDebugMarker&) = delete;
+		ScopedDebugMarker(ScopedDebugMarker&&) = default;
 		~ScopedDebugMarker();
+
+		ScopedDebugMarker& operator=(const ScopedDebugMarker&) = delete;
+	};
+
+	class Query final
+	{
+	public:
+		Query(GLenum type);
+		Query(const Query&) = delete;
+		Query(Query&&) = default;
+		~Query();
+
+		Query& operator=(const Query&) = delete;
+
+		void Begin();
+		void End();
+
+	private:
+		GLenum m_type;
+		GLuint m_id;
 	};
 } // namespace gl4A